+++ src/Answer.King.Domain/Orders/Product.cs
@@ -4,7 +4,7 @@ namespace Answer.King.Domain.Orders
 {
     public class Product
     {
-        public Product(Guid id, string name, string description, Category category, float price)
+        public Product(Guid id, string name, string description, Category category, double price)
         {
             Guard.AgainstDefaultValue(nameof(id), id);
             Guard.AgainstNullOrEmptyArgument(nameof(name), name);
@@ -25,7 +25,7 @@ namespace Answer.King.Domain.Orders
 
         public string Description { get; }
 
-        public float Price { get; }
+        public double Price { get; }
 
         public Category Category { get; }
     }
